+++ src/layout/navBars/breadcrumb/user.vue
@@ -65,6 +65,7 @@ import screenfull from 'screenfull';
 import { useI18n } from 'vue-i18n';
 import { resetRoute } from '/@/router/index';
 import { useStore } from '/@/store/index';
+import { useTitle } from '/@/utils/setWebTitle';
 import { Session, Local } from '/@/utils/storage';
 import Search from '/@/layout/navBars/breadcrumb/search.vue';
 export default {
@@ -75,6 +76,7 @@ export default {
 		const { proxy } = getCurrentInstance() as any;
 		const router = useRouter();
 		const store = useStore();
+    const title = useTitle();
 		const searchRef = ref();
 		const state = reactive({
 			isScreenfull: false,
@@ -104,7 +106,10 @@ export default {
 				return false;
 			}
 			screenfull.toggle();
-			state.isScreenfull = !state.isScreenfull;
+      screenfull.on('change', () => {
+        if (screenfull.isFullscreen) state.isScreenfull = true;
+        else state.isScreenfull = false;
+      });
 		};
 		// 布局配置 icon 点击时
 		const onLayoutSetingClick = () => {
@@ -171,6 +176,7 @@ export default {
 			Local.set('themeConfig', getThemeConfig.value);
 			proxy.$i18n.locale = lang;
 			initI18n();
+      title();
 		};
 		// 初始化言语国际化
 		const initI18n = () => {