+++ src/utils/getStyleSheets.ts
@@ -36,13 +36,15 @@ const getElementPlusIconfont = () => {
 			for (let i = 0; i < styles.length; i++) {
 				for (let j = 0; j < styles[i].cssRules.length; j++) {
 					if (styles[i].cssRules[j].selectorText && styles[i].cssRules[j].selectorText.indexOf('.el-icon-') === 0) {
-						sheetsIconList.push(
-							`${styles[i].cssRules[j].selectorText.substring(1, styles[i].cssRules[j].selectorText.length).replace(/\:\:before/gi, '')}`
-						);
+						if (/::before/.test(styles[i].cssRules[j].selectorText)) {
+							sheetsIconList.push(
+								`${styles[i].cssRules[j].selectorText.substring(1, styles[i].cssRules[j].selectorText.length).replace(/\:\:before/gi, '')}`
+							);
+						}
 					}
 				}
 			}
-			if (sheetsIconList.length > 0) resolve(sheetsIconList);
+			if (sheetsIconList.length > 0) resolve(sheetsIconList.reverse());
 			else reject('未获取到值，请刷新重试');
 		});
 	});
@@ -67,13 +69,15 @@ const getAwesomeIconfont = () => {
 						sheetsList[i].cssRules[j].selectorText.indexOf('.fa-') === 0 &&
 						sheetsList[i].cssRules[j].selectorText.indexOf(',') === -1
 					) {
-						sheetsIconList.push(
-							`${sheetsList[i].cssRules[j].selectorText.substring(1, sheetsList[i].cssRules[j].selectorText.length).replace(/\:\:before/gi, '')}`
-						);
+						if (/::before/.test(sheetsList[i].cssRules[j].selectorText)) {
+							sheetsIconList.push(
+								`${sheetsList[i].cssRules[j].selectorText.substring(1, sheetsList[i].cssRules[j].selectorText.length).replace(/\:\:before/gi, '')}`
+							);
+						}
 					}
 				}
 			}
-			if (sheetsIconList.length > 0) resolve(sheetsIconList);
+			if (sheetsIconList.length > 0) resolve(sheetsIconList.reverse());
 			else reject('未获取到值，请刷新重试');
 		});
 	});